2dcontext/drawing-text-to-the-canvas/2d.text.measure.offscreencanvas.fontBoundingBox.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.text.measure.fontBoundingBox</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("/fonts/CanvasTest.ttf");
}
</style>
<body class="show_output">

<h1>2d.text.measure.offscreencanvas.fontBoundingBox</h1>
<p class="desc">Testing fontBoundingBox for OffscreenCanvas</p>

<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>

<ul id="d"></ul>
<script>
var t = async_test("Testing fontBoundingBox for OffscreenCanvas");
test(function () {
  deferTest();
  var f = new FontFace("CanvasTest", "/fonts/CanvasTest.ttf");
  document.fonts.add(f);
  document.fonts.ready.then(() => {
    step_timeout(t.step_func_done(function () {
      var offscreen = new OffscreenCanvas(100, 50);
      var ctx = offscreen.getContext('2d');
      ctx.font = '50px CanvasTest';
      ctx.direction = 'ltr';
      ctx.align = 'left'
      ctx.baseline = 'alphabetic'
      _assertSame(ctx.measureText('A').fontBoundingBoxAscent, 85, "ctx.measureText('A').fontBoundingBoxAscent", "85");
      _assertSame(ctx.measureText('A').fontBoundingBoxDescent, 39, "ctx.measureText('A').fontBoundingBoxDescent", "39");

      _assertSame(ctx.measureText('ABCD').fontBoundingBoxAscent, 85, "ctx.measureText('ABCD').fontBoundingBoxAscent", "85");
      _assertSame(ctx.measureText('ABCD').fontBoundingBoxDescent, 39, "ctx.measureText('ABCD').fontBoundingBoxDescent", "39");
    }), 0);
  });
});
</script>
</body>>
